pyproject.toml
[tool.poetry]
name = "bybit-trading-bot"
version = "1.0.0"
description = "Automated Bybit Spot Trading Bot with AI Analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12.6"
pybit = "^5.8.0"
python-telegram-bot = "^21.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
websockets = "^12.0"
apscheduler = "^3.10.4"
pytz = "^2024.1"
matplotlib = "^3.8.0"
pandas = "^2.1.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
